chore: take 기능 수정
- take 관련 컨트랙트 연결 수정
- 로직 뷰 분리
